Add freezer and combine exception tests
